# investment-system/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 投资管理系统统一接口契约
  version: 1.1.0
  description: |
    LTC 投资管理系统正式接口契约。认证使用 HttpOnly Cookie，会话写操作同时校验 X-CSRF-Token。
    所有投资业务单据统一按“经办人、经办部门负责人、经办部门分管领导”顺序流转。
servers:
  - url: http://127.0.0.1:3000/api/v1
    description: 本地及部署环境 API 基地址
security:
  - cookieAuth: []
...
paths:
...
  /investment-intentions:
    get:
      operationId: listInvestmentIntentions
      summary: 分页查询我的发起、我的待办或全部投资意向
      x-permissions: [investment.intention.read, investment.intention.read_all]
      parameters:
        - $ref: '#/components/parameters/Mode'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses: *okResponses
    post:
      operationId: createInvestmentIntention
      summary: 创建股权投资意向待发单
      x-permissions: [investment.intention.create]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntentionInput'
      responses: *createdResponses
  /investment-intentions/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      operationId: getInvestmentIntention
      summary: 查询有权访问的投资意向详情
      x-permissions: [investment.intention.read, investment.intention.read_all]
      responses: *okResponses
    put:
      operationId: updateInvestmentIntention
      summary: 修改本人待发或退回的投资意向
      x-permissions: [investment.intention.update]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntentionUpdateInput'
      responses: *okResponses
  /investment-intentions/{id}/submit:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    post:
      operationId: submitInvestmentIntention
      summary: 提交投资意向至经办部门负责人
      x-permissions: [investment.intention.submit]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      responses: *okResponses
  /investment-intentions/{id}/approve:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    post:
      operationId: approveInvestmentIntention
      summary: 指定部门负责人或分管领导同意投资意向
      x-permissions: [investment.intention.approve_department, investment.intention.approve_supervising]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewInput'
      responses: *okResponses
  /investment-intentions/{id}/return:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    post:
      operationId: returnInvestmentIntention
      summary: 指定部门负责人或分管领导退回投资意向
      x-permissions: [investment.intention.approve_department, investment.intention.approve_supervising]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewInput'
      responses: *okResponses
  /investment-intentions/{id}/attachments:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    post:
      operationId: uploadInvestmentIntentionAttachment
      summary: 为本人可编辑的投资意向上传安全校验后的附件
      x-permissions: [investment.attachment.manage]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses: *createdResponses
  /attachments/{id}/download:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      operationId: downloadInvestmentAttachment
      summary: 下载有业务查看权限的投资意向附件
      x-permissions: [investment.attachment.read]
      responses: *okResponses
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: ltc_session
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    CsrfHeader:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string, minLength: 64, maxLength: 64 }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Mode:
      name: mode
      in: query
      schema: { type: string, enum: [mine, todo, all], default: mine }
...
  schemas:
...
    ReviewInput:
      type: object
      additionalProperties: false
      properties:
        comment: { type: [string, 'null'], maxLength: 2000 }
    IntentionInput:
      type: object
      additionalProperties: false
      properties:
        investmentEntityId: { type: [string, 'null'], format: uuid }
        applicationDate: { type: [string, 'null'], format: date }
        projectName: { type: [string, 'null'], maxLength: 300 }
        investmentMethod: { type: [string, 'null'], maxLength: 32 }
        majorProject: { type: [boolean, 'null'] }
        plannedStartDate: { type: [string, 'null'], format: date }
        plannedEndDate: { type: [string, 'null'], format: date }
        projectLeaderUserId: { type: [string, 'null'], format: uuid }
        contactPhone: { type: [string, 'null'], maxLength: 50 }
        projectLocation: { type: [string, 'null'], maxLength: 500 }
        projectSummary: { type: [string, 'null'], maxLength: 65535 }
        mainContent: { type: [string, 'null'], maxLength: 65535 }
        targetCompanyId: { type: [string, 'null'], maxLength: 64 }
        targetCompanyName: { type: [string, 'null'], maxLength: 300 }
        mainBusiness: { type: [string, 'null'], maxLength: 1000 }
        investmentDirection: { type: [string, 'null'], maxLength: 64 }
        domesticOverseas: { type: [string, 'null'], enum: [DOMESTIC, OVERSEAS, null] }
        currencyCode: { type: [string, 'null'], maxLength: 8 }
        exchangeRate: { type: [string, number, 'null'] }
        plannedShareholdingRatio: { type: [string, number, 'null'] }
        projectTotalInvestment: { type: [string, number, 'null'] }
        plannedInvestment: { type: [string, number, 'null'] }
        expectedReturnRate: { type: [string, number, 'null'] }
    IntentionUpdateInput:
      allOf:
        - $ref: '#/components/schemas/IntentionInput'
        - type: object
          required: [version]
          properties:
            version: { type: integer, minimum: 1 }
...